Update composer libraries for security reasons
The commit contains multiple patches in the composer.json and composer.lock files.
The changes have updated the versions of many libraries to address vulnerabilities and bugs, improving the project's stability and security. The updates were not limited to Laravel packages and also touch on laminas, psalm, phpunit library among others. The version updates will enforce the usage of the newer, secure versions when doing a composer install/update.